+++ packages/shiki_flutter/lib/shiki_flutter.dart
@@ -39,8 +39,9 @@ export 'src/onig/onig.dart'
         OnigString,
         OnigMatch,
         OnigCaptureIndex;
-export 'src/flutter/code_list_view.dart' show ShikiCodeListView, GutterStyle;
+export 'src/flutter/code_list_view.dart' show ShikiCodeListView;
 export 'src/flutter/code_view.dart' show ShikiCodeView;
+export 'src/flutter/gutter.dart' show GutterStyle;
 export 'src/flutter/render.dart'
     show
         codeToTextSpan,